requirements.txt
asgiref==3.10.0
crispy-bootstrap4==2025.6
Django==5.2.7
django-crispy-forms==2.4
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
